# docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: Order System API
  description: API for creating, viewing and cancelling orders in the Order System.
  version: 1.0.0
  license:
    name: MIT

servers:
  - url: http://localhost:8083
    description: Local development server

tags:
  - name: Products
    description: Product catalog used when creating orders.
  - name: Orders
    description: Order creation, retrieval and cancellation.

security: []

paths:
  /api/products:
    get:
      tags: [Products]
      operationId: listProducts
      security: []
      summary: List available products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      security: []
      summary: Create an order
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Client-supplied key to make order creation idempotent.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product or inventory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Insufficient inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrder
      security: []
      summary: Get an order by id
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      operationId: cancelOrder
      security: []
      summary: Cancel an order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Order cannot be cancelled in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: The order id.
      schema:
        type: string
        format: uuid

  schemas:
    Product:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
          format: double

    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'

    CreateOrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1

    Order:
      type: object
      required: [id, status, items, totalAmount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderItem:
      type: object
      required: [productId, productName, unitPrice, quantity, lineTotal]
      properties:
        productId:
          type: string
          format: uuid
        productName:
          type: string
        unitPrice:
          type: number
          format: double
        quantity:
          type: integer
        lineTotal:
          type: number
          format: double

    OrderStatus:
      type: string
      enum:
        - PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELLED

    Error:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time